+++ packages/sqlite-datasource/src/index.ts
@@ -1,2 +1,6 @@
-export { NodeSqliteTelemetryDatasource } from "./datasource.js";
+export {
+  OptimizedDatasource,
+  createOptimizedDatasource,
+} from "./optimized-datasource.js";
+export { DbDatasource } from "./db-datasource.js";
 export { initializeDatabase } from "./initialize-database.js";